requirements.txt
numpy>=1.26.4
pandas>=2.2.3
matplotlib>=3.9.2
seaborn>=0.13.2
scikit-learn>=1.5.1
streamlit>=1.38.0
streamlit-folium>=0.15.0
python-dotenv>=1.0.1
twilio>=8.0.0
